requirements.txt
transformers
timm
datasets
wandb
lm-eval
torchao
sentencepiece
tiktoken[blobfile]
bitsandbytes
torchdata
